requirements.txt
mkdocs-glightbox
mkdocs-material
mkdocs-macros-plugin
aiohttp
aiosqlite
python-dotenv
